refactor: 컴포넌트 분리/합성 및 레이아웃 수정
