+++ src/pages/AddCard/index.jsx
@@ -3,6 +3,8 @@ import { useState } from 'react';
 import AddCardForm from '../../components/AddCardForm';
 import Modal from '../../components/Modal';
 import Palette from '../../components/Palette';
+import Header from '../../components/Header';
+import FallbackButton from '../../components/FallbackButton';
 
 const AddCard = () => {
   const [isModalOpened, setModalStatus] = useState(false);
@@ -17,6 +19,12 @@ const AddCard = () => {
   return (
     <CardFormProvider>
       <>
+        <Header>
+          <>
+            <FallbackButton />
+            <p>카드추가</p>
+          </>
+        </Header>
         <AddCardForm openModal={openModal} />
         <Modal isModalOpened={isModalOpened} closeModal={closeModal}>
           <Palette closeModal={closeModal} />
